[TM-1425] Add delayed job to run analysis
